Steam added, change of settings
steam particles have been added, water will now turn into steam at temperatures of 100 or greater
steam will turn back into water at temperatures of 98 or less

settings to do with health/weight/thermal conductivity have been changed to an array that is initialized with settings for each particle

+++ Code Test/Code/EntityHandle.h
@@ -4,11 +4,11 @@ void EntityHandle()
 {
 	if (!pauseParticles)
 	{
-		for (int i = 0; i < particleList.size(); i++)
+		for (size_t i = 0; i < particleList.size(); i++)
 			particleList.at(i)->HandleEvents();
 
 		//handling entities
-		for (int i = 0; i < particleList.size(); i++)
+		for (size_t i = 0; i < particleList.size(); i++)
 			particleList.at(i)->HandlePhysics();
 	}
 }
